backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.32
httpx==0.27.2
python-dotenv==1.0.1
